// frontend/packages/api/models/did-event-historical-count.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OptiPrism
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * Contact: api@optiprism.io
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { PropertyFilterOperation } from './property-filter-operation';
import { TimeAfterFirstUse } from './time-after-first-use';
import { TimeBetween } from './time-between';
import { TimeLast } from './time-last';
import { TimeWindowEach } from './time-window-each';
/**
 * find users with Nth event occurance
 * @export
 * @interface DidEventHistoricalCount
 */
export interface DidEventHistoricalCount {
    /**
     * 
     * @type {string}
     * @memberof DidEventHistoricalCount
     */
    type: DidEventHistoricalCountTypeEnum;
    /**
     * 
     * @type {PropertyFilterOperation}
     * @memberof DidEventHistoricalCount
     */
    operation: PropertyFilterOperation;
    /**
     * 
     * @type {number}
     * @memberof DidEventHistoricalCount
     */
    value?: number;
    /**
     * 
     * @type {TimeBetween | TimeLast | TimeAfterFirstUse | TimeWindowEach}
     * @memberof DidEventHistoricalCount
     */
    time: TimeBetween | TimeLast | TimeAfterFirstUse | TimeWindowEach;
}

/**
    * @export
    * @enum {string}
    */
export enum DidEventHistoricalCountTypeEnum {
    HistoricalCount = 'historicalCount'
}